Implement graph iterators
